+++ adetailer/args.py
@@ -5,7 +5,6 @@ from dataclasses import dataclass
 from functools import cached_property, partial
 from typing import Any, Literal, NamedTuple, Optional
 
-import pydantic
 from pydantic import (
     BaseModel,
     Extra,
@@ -14,7 +13,6 @@ from pydantic import (
     PositiveInt,
     confloat,
     conint,
-    constr,
     validator,
 )
 
@@ -34,16 +32,17 @@ class Arg(NamedTuple):
 
 class ArgsList(UserList):
     @cached_property
-    def attrs(self) -> tuple[str]:
+    def attrs(self) -> tuple[str, ...]:
         return tuple(attr for attr, _ in self)
 
     @cached_property
-    def names(self) -> tuple[str]:
+    def names(self) -> tuple[str, ...]:
         return tuple(name for _, name in self)
 
 
 class ADetailerArgs(BaseModel, extra=Extra.forbid):
     ad_model: str = "None"
+    ad_model_classes: str = ""
     ad_prompt: str = ""
     ad_negative_prompt: str = ""
     ad_confidence: confloat(ge=0.0, le=1.0) = 0.3
@@ -113,6 +112,7 @@ class ADetailerArgs(BaseModel, extra=Extra.forbid):
         p = {name: getattr(self, attr) for attr, name in ALL_ARGS}
         ppop = partial(self.ppop, p)
 
+        ppop("ADetailer model classes")
         ppop("ADetailer prompt")
         ppop("ADetailer negative prompt")
         ppop("ADetailer mask only top k largest", cond=0)
@@ -185,6 +185,7 @@ class ADetailerArgs(BaseModel, extra=Extra.forbid):
 
 _all_args = [
     ("ad_model", "ADetailer model"),
+    ("ad_model_classes", "ADetailer model classes"),
     ("ad_prompt", "ADetailer prompt"),
     ("ad_negative_prompt", "ADetailer negative prompt"),
     ("ad_confidence", "ADetailer confidence"),